+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - The "age indicator" label now updates once per minute to a rounded-down value of the last modified time
+- The study session due date updates with the passage of time and with app foregrounding
 
 ## [0.21.0] - 2020-03-07